Widen extra_flags to unsigned long long, add ITEM_BONDED flag
extra_flags on OBJ_DATA and OBJ_INDEX_DATA was `int` (32-bit) but
already used BIT_32 (ITEM_TWO_HANDED), which overflows signed int.
Widen to `unsigned long long` to properly support all existing flags
and make room for new ones.

Changes:
- Widen extra_flags field in both OBJ_DATA and OBJ_INDEX_DATA structs
- Add ITEM_BONDED (BIT_33) for the weapon bonding system
- Update save/load: fprintf %d -> %llu, fread_number -> fread_number_ull
- Update extra_bit_name() signature to unsigned long long
- Widen build_setobject value variable to unsigned long long
- Add "bonded" to tab_obj_flags and extra_bit_name output
- Replace value[9] marker hack with proper ITEM_BONDED flag
- Fix IS_OBJ_BONDED macro with !! to avoid bool truncation
  (unsigned char bool truncates high bits of unsigned long long)

+++ src/headers/ack.h
@@ -898,7 +898,7 @@ struct obj_index_data
    char *description;
    unsigned short vnum;
    int item_type;
-   int extra_flags;
+   unsigned long long extra_flags;
    int wear_flags;
    /*
     * class_flags changed to item_apply for magic apply, etc
@@ -946,7 +946,7 @@ struct obj_data
    char *short_descr;
    char *description;
    int item_type;
-   int extra_flags;
+   unsigned long long extra_flags;
    int wear_flags;
    int item_apply;
    int wear_loc;
